requirements.txt
streamlit>=1.32.0
google-cloud-bigquery>=3.20.0
pandas>=2.0.0
plotly>=5.18.0
db-dtypes>=1.2.0
